+++ src/main/java/com/interior/adapter/inbound/company/CompanyController.java
@@ -2,8 +2,8 @@ package com.interior.adapter.inbound.company;
 
 import com.interior.adapter.inbound.company.webdto.CreateCompanyDto;
 import com.interior.adapter.inbound.company.webdto.GetCompanyDto.GetCompanyResDto;
-import com.interior.application.command.company.CompanyCommandService;
-import com.interior.application.command.company.dto.CreateCompanyServiceDto;
+import com.interior.application.commands.company.CompanyCommandService;
+import com.interior.application.commands.company.dto.CreateCompanyServiceDto;
 import com.interior.application.query.company.CompanyQueryService;
 import com.interior.domain.user.User;
 import jakarta.validation.Valid;
@@ -11,7 +11,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
