+++ src/main/java/com/interior/adapter/inbound/user/UserController.java
@@ -4,7 +4,7 @@ import com.interior.adapter.inbound.user.webdto.LoadUserDto.LoadUserResDto;
 import com.interior.adapter.inbound.user.webdto.RequestValidation;
 import com.interior.adapter.inbound.user.webdto.SignUpDto;
 import com.interior.adapter.inbound.user.webdto.SignUpDto.SignUpResDto;
-import com.interior.application.command.user.UserCommandService;
+import com.interior.application.commands.user.UserCommandService;
 import com.interior.application.query.user.UserQueryService;
 import com.interior.domain.user.User;
 import com.interior.domain.util.BoolType;
