+++ src/main/java/com/interior/adapter/outbound/email/MailSender.java
@@ -2,7 +2,7 @@ package com.interior.adapter.outbound.email;
 
 import static com.interior.adapter.common.exception.ErrorType.UNABLE_TO_SEND_EMAIL;
 
-import com.interior.application.command.util.email.template.EmailTemplate;
+import com.interior.application.commands.util.email.template.EmailTemplate;
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.MimeMessage;
 import lombok.RequiredArgsConstructor;
