--- src/main/java/com/interior/application/command/business/dto/CreateBusinessServiceDto.java
@@ -1,19 +1,0 @@
-package com.interior.application.command.business.dto;
-
-import java.math.BigDecimal;
-
-public class CreateBusinessServiceDto {
-
-    public record CreateBusinessMaterialDto(
-            String name,
-            String usageCategory,
-            String category,
-            BigDecimal amount,
-            String unit,
-            String memo,
-            String materialCostPerUnit,
-            String laborCostPerUnit
-    ) {
-
-    }
-}
