--- src/main/java/com/interior/application/command/business/dto/ReviseBusinessServiceDto.java
@@ -1,8 +1,0 @@
-package com.interior.application.command.business.dto;
-
-public class ReviseBusinessServiceDto {
-
-    public record Req(String changeBusinessName) {
-
-    }
-}
