--- src/main/java/com/interior/application/command/company/dto/CreateCompanyServiceDto.java
@@ -1,15 +1,0 @@
-package com.interior.application.command.company.dto;
-
-public class CreateCompanyServiceDto {
-
-    public record CreateCompanyDto(
-            String companyName,
-            String zipCode,
-            String mainAddress,
-            String subAddress,
-            String bdgNumber,
-            String tel
-    ) {
-
-    }
-}
