--- src/main/java/com/interior/application/command/log/business/dto/event/BusinessDeleteLogEvent.java
@@ -1,5 +1,0 @@
-package com.interior.application.command.log.business.dto.event;
-
-public record BusinessDeleteLogEvent(Long businessId, Long updaterId, String originalBusinessName) {
-
-}
