--- src/main/java/com/interior/application/command/log/business/material/dto/event/BusinessMaterialCreateLogEvent.java
@@ -1,6 +1,0 @@
-package com.interior.application.command.log.business.material.dto.event;
-
-public record BusinessMaterialCreateLogEvent(Long businessId, Long businessMaterialId,
-                                             Long updaterId, String afterData) {
-
-}
