--- src/main/java/com/interior/application/command/log/business/material/dto/event/BusinessMaterialDeleteLogEvent.java
@@ -1,5 +1,0 @@
-package com.interior.application.command.log.business.material.dto.event;
-
-public record BusinessMaterialDeleteLogEvent(Long businessId, Long materialId, Long updaterId) {
-
-}
