--- src/main/java/com/interior/application/command/log/sms/event/SmsSendResultLogEvent.java
@@ -1,11 +1,0 @@
-package com.interior.application.command.log.sms.event;
-
-import java.util.HashMap;
-
-public record SmsSendResultLogEvent(
-        String senderPhoneNumber,
-        String to,
-        HashMap<String, String> msg,
-        String platformType) {
-
-}
