--- src/main/java/com/interior/application/command/util/email/template/EmailTemplate.java
@@ -1,10 +1,0 @@
-package com.interior.application.command.util.email.template;
-
-public interface EmailTemplate {
-
-    String getTitle();
-
-    String getContent();
-
-    String getTargetEmail();
-}
