+++ src/main/java/com/interior/application/commands/business/dto/CreateBusinessServiceDto.java
@@ -1,0 +1,19 @@
+package com.interior.application.commands.business.dto;
+
+import java.math.BigDecimal;
+
+public class CreateBusinessServiceDto {
+
+    public record CreateBusinessMaterialDto(
+            String name,
+            String usageCategory,
+            String category,
+            BigDecimal amount,
+            String unit,
+            String memo,
+            String materialCostPerUnit,
+            String laborCostPerUnit
+    ) {
+
+    }
+}
