+++ src/main/java/com/interior/application/commands/business/dto/ReviseBusinessServiceDto.java
@@ -1,0 +1,8 @@
+package com.interior.application.commands.business.dto;
+
+public class ReviseBusinessServiceDto {
+
+    public record Req(String changeBusinessName) {
+
+    }
+}
