+++ src/main/java/com/interior/application/commands/company/dto/CreateCompanyServiceDto.java
@@ -1,0 +1,15 @@
+package com.interior.application.commands.company.dto;
+
+public class CreateCompanyServiceDto {
+
+    public record CreateCompanyDto(
+            String companyName,
+            String zipCode,
+            String mainAddress,
+            String subAddress,
+            String bdgNumber,
+            String tel
+    ) {
+
+    }
+}
