+++ src/main/java/com/interior/application/commands/log/business/dto/event/BusinessDeleteLogEvent.java
@@ -1,0 +1,5 @@
+package com.interior.application.commands.log.business.dto.event;
+
+public record BusinessDeleteLogEvent(Long businessId, Long updaterId, String originalBusinessName) {
+
+}
