+++ src/main/java/com/interior/application/commands/log/business/material/dto/event/BusinessMaterialCreateLogEvent.java
@@ -1,0 +1,6 @@
+package com.interior.application.commands.log.business.material.dto.event;
+
+public record BusinessMaterialCreateLogEvent(Long businessId, Long businessMaterialId,
+                                             Long updaterId, String afterData) {
+
+}
