+++ src/main/java/com/interior/application/commands/log/business/material/dto/event/BusinessMaterialDeleteLogEvent.java
@@ -1,0 +1,5 @@
+package com.interior.application.commands.log.business.material.dto.event;
+
+public record BusinessMaterialDeleteLogEvent(Long businessId, Long materialId, Long updaterId) {
+
+}
