+++ src/main/java/com/interior/application/commands/log/sms/SmsSendResultLogService.java
@@ -1,0 +1,33 @@
+package com.interior.application.commands.log.sms;
+
+import com.interior.application.commands.log.sms.event.SmsSendResultLogEvent;
+import com.interior.domain.sms.SmsSendResult;
+import com.interior.domain.sms.repository.SmsRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.event.EventListener;
+import org.springframework.scheduling.annotation.Async;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Async
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class SmsSendResultLogService {
+
+    private final SmsRepository smsRepository;
+
+    @EventListener
+    @Transactional
+    public void createSmsSendResultLog(final SmsSendResultLogEvent event) {
+        smsRepository.save(
+                SmsSendResult.of(
+                        event.senderPhoneNumber(),
+                        event.to(),
+                        event.msg(),
+                        event.platformType()
+                )
+        );
+    }
+}
