+++ src/main/java/com/interior/application/commands/log/sms/event/SmsSendResultLogEvent.java
@@ -1,0 +1,11 @@
+package com.interior.application.commands.log.sms.event;
+
+import java.util.HashMap;
+
+public record SmsSendResultLogEvent(
+        String senderPhoneNumber,
+        String to,
+        HashMap<String, String> msg,
+        String platformType) {
+
+}
