+++ src/main/java/com/interior/application/commands/util/email/template/EmailTemplate.java
@@ -1,0 +1,10 @@
+package com.interior.application.commands.util.email.template;
+
+public interface EmailTemplate {
+
+    String getTitle();
+
+    String getContent();
+
+    String getTargetEmail();
+}
