+++ src/main/java/com/interior/domain/business/repository/BusinessRepository.java
@@ -2,7 +2,7 @@ package com.interior.domain.business.repository;
 
 import com.interior.adapter.inbound.business.enumtypes.QueryType;
 import com.interior.adapter.outbound.jpa.repository.business.dto.ReviseBusinessMaterial;
-import com.interior.application.command.business.dto.ReviseBusinessServiceDto;
+import com.interior.application.commands.business.dto.ReviseBusinessServiceDto;
 import com.interior.domain.business.Business;
 import com.interior.domain.business.log.BusinessLog;
 import com.interior.domain.business.material.BusinessMaterial;
@@ -10,7 +10,6 @@ import com.interior.domain.business.material.log.BusinessMaterialLog;
 import com.interior.domain.business.progress.ProgressType;
 import com.interior.domain.business.repository.dto.CreateBusiness;
 import com.interior.domain.business.repository.dto.CreateBusinessMaterial;
-
 import java.util.List;
 
 public interface BusinessRepository {
@@ -32,7 +31,7 @@ public interface BusinessRepository {
     boolean deleteBusiness(final Long companyId, final Long businessId);
 
     boolean reviseBusiness(final Long userId, final Long businessId,
-                           final ReviseBusinessServiceDto.Req req);
+            final ReviseBusinessServiceDto.Req req);
 
     boolean reviseUsageCategoryOfMaterial(
             final Long businessId,
